requirements.txt
psycopg2 ~= 2.7.4
Django ~= 2.0.5
django-rest-multiple-models ~= 2.1.0
djangorestframework ~= 3.8.2
django-cors-headers ~= 2.2.0
